chore: Update home page content and layout
The commit updates the content and layout of the home page. It adds a welcome message and description, as well as a list of current features. It also updates the buttons for registration, login, and accessing the family area. These changes aim to improve the user experience and provide clearer information to visitors.

Co-authored-by: Leanderson Nunes da Rocha <Leanderson01@users.noreply.github.com>