chore: Update publicRoutes in middleware.ts
This commit updates the publicRoutes configuration in the middleware.ts file. The previous implementation used a function to determine public routes based on the URL pathname, but it has been changed to use a regular expression pattern ['((?!^/club).*)']. This change ensures that locale specific sign-in pages are public, except for those under the '/club' path.

Co-authored-by: Leanderson Nunes da Rocha <Leanderson01@users.noreply.github.com>